+++ lib/ecache.dart
@@ -10,6 +10,7 @@ export 'src/cache/expiration_cache.dart';
 export 'src/cache/lfu_cache.dart';
 export 'src/cache/lru_cache.dart';
 export 'src/cache/simple_cache.dart';
+export 'src/cache/sync_cache.dart';
 export 'src/cache_entry.dart';
 export 'src/storage.dart';
 export 'src/storage/simple_storage.dart';