+++ lib/src/cache/sync_cache.dart
@@ -2,19 +2,19 @@ import 'dart:async';
 
 import '../../ecache.dart';
 
-/// A generic class for [Cache] implementations.
+/// A generic, stripped down class for [Cache] implementations if you do not need async calls.
 ///
 /// This class provides the core caching logic, delegating entry management and
 /// eviction policies to a specified [AbstractStrategy]. It uses a [Storage]
 /// mechanism to hold the cache entries.
-class DefaultCache<K, V> extends Cache<K, V> {
+class SyncCache<K, V> extends Cache<K, V> {
   @override
   final Storage<K, V> storage;
 
   /// The strategy used for cache entry management and eviction.
   final AbstractStrategy<K, V> strategy;
 
-  /// Creates a new [DefaultCache].
+  /// Creates a new [SyncCache].
   ///
   /// A [capacity] for the cache must be provided.
   ///
@@ -23,18 +23,16 @@ class DefaultCache<K, V> extends Cache<K, V> {
   ///
   /// An optional [strategy] can be provided. If not, a [SimpleStrategy]
   /// instance is used.
-  DefaultCache({Storage<K, V>? storage, required int capacity, AbstractStrategy<K, V>? strategy})
+  SyncCache({Storage<K, V>? storage, required int capacity, AbstractStrategy<K, V>? strategy})
       : storage = storage ?? SimpleStorage<K, V>(),
         strategy = strategy ?? SimpleStrategy<K, V>() {
     this.strategy.init(this.storage, capacity);
   }
 
+  /// Synchronously returns the element for the given [key], or `null` if the key is not found.
   @override
   V? get(K key) {
     CacheEntry<K, V>? entry = strategy.get(key);
-    if (entry is ProducerCacheEntry<K, V>) {
-      throw Exception("Cannot get a value from a producer since the value is a future and the get() method is synchronously");
-    }
     return entry?.value;
   }
 
@@ -44,70 +42,23 @@ class DefaultCache<K, V> extends Cache<K, V> {
   /// that will complete with the produced value.
   @override
   Future<V?> getAsync(K key) async {
-    CacheEntry<K, V>? entry = strategy.get(key);
-    if (entry is ProducerCacheEntry<K, V>) {
-      return entry.completer.future;
-    }
-    return entry?.value;
+    throw UnimplementedError();
   }
 
   @override
   Future<V> getOrProduce(K key, Produce<K, V> produce, [int timeoutMilliseconds = 60000]) async {
-    CacheEntry<K, V>? entry = storage.get(key);
-    if (entry != null) {
-      if (entry is ProducerCacheEntry<K, V>) {
-        return entry.completer.future;
-      }
-      return entry.value!;
-    }
-
-    strategy.onCapacity(key);
-    ProducerCacheEntry<K, V> producer = strategy.createProducerCacheEntry(key, produce);
-    storage.set(key, producer);
-    unawaited(producer.start(key, timeoutMilliseconds));
-
-    try {
-      V value = await producer.completer.future;
-      storage.set(key, strategy.createCacheEntry(key, value));
-      return value;
-    } catch (e) {
-      storage.remove(key);
-      rethrow;
-    }
+    throw UnimplementedError();
   }
 
   @override
   Future<V> produce(K key, Produce<K, V> produce, [int timeoutMilliseconds = 60000]) async {
-    CacheEntry<K, V>? entry = storage.get(key);
-    if (entry != null) {
-      if (entry is ProducerCacheEntry<K, V>) {
-        return entry.completer.future;
-      }
-    }
-
-    strategy.onCapacity(key);
-    ProducerCacheEntry<K, V> producer = strategy.createProducerCacheEntry(key, produce);
-    storage.set(key, producer);
-    unawaited(producer.start(key, timeoutMilliseconds));
-
-    try {
-      V value = await producer.completer.future;
-      // replace this entry with a normal cacheEntry so that get() works
-      storage.set(key, strategy.createCacheEntry(key, value));
-      return value;
-    } catch (e) {
-      storage.remove(key);
-      rethrow;
-    }
+    throw UnimplementedError();
   }
 
   @override
   V getOrProduceSync(K key, ProduceSync<K, V> produce) {
     CacheEntry<K, V>? entry = storage.get(key);
     if (entry != null) {
-      if (entry is ProducerCacheEntry<K, V>) {
-        throw Exception("Cannot get a value from a producer since the value is a future and the get() method is synchronously");
-      }
       return entry.value!;
     }